Feat: 도메인 주도 개발 적용
User 서비스

+++ src/main/java/com/hhplus/commerce/spring/application/user/dto/UserInfo.java
@@ -3,10 +3,12 @@ package com.hhplus.commerce.spring.application.user.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class UserInfo {
     private Long id;
     private String name;
-    private int point;
+    private BigDecimal point;
 }
