+++ src/main/java/com/hhplus/commerce/spring/domain/user/dto/UserCommand.java
@@ -3,12 +3,14 @@ package com.hhplus.commerce.spring.domain.user.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 public class UserCommand {
 
     @Getter
     @Builder
     public static class PointCharge {
         private Long userId;
-        private Integer chargePoint;
+        private BigDecimal chargePoint;
     }
 }
