+++ src/main/java/com/hhplus/commerce/spring/old/api/user/controller/response/BalanceChargeResponse.java
@@ -3,10 +3,12 @@ package com.hhplus.commerce.spring.old.api.user.controller.response;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class BalanceChargeResponse {
 
     private Long userId;
-    private int userPoint;
+    private BigDecimal userPoint;
 }
