+++ src/main/java/com/hhplus/commerce/spring/old/api/user/model/User.java
@@ -3,10 +3,12 @@ package com.hhplus.commerce.spring.old.api.user.model;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class User {
     private Long id;
     private String name;
-    private Integer point;
+    private BigDecimal point;
 }
