+++ src/main/java/com/hhplus/commerce/spring/presentation/user/dto/response/UserResponseDTO.java
@@ -3,10 +3,12 @@ package com.hhplus.commerce.spring.presentation.user.dto.response;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class UserResponseDTO {
     private Long id;
     private String name;
-    private int point;
+    private BigDecimal point;
 }
