+++ src/main/java/com/hhplus/commerce/spring/presentation/user/mapper/UserDTORequestMapper.java
@@ -2,9 +2,11 @@ package com.hhplus.commerce.spring.presentation.user.mapper;
 
 import com.hhplus.commerce.spring.application.user.dto.request.UserPointChargeRequest;
 
+import java.math.BigDecimal;
+
 public class UserDTORequestMapper {
 
-    public static UserPointChargeRequest toUserPointChargeRequest(Long userId, int chargePoint) {
+    public static UserPointChargeRequest toUserPointChargeRequest(Long userId, BigDecimal chargePoint) {
         return UserPointChargeRequest.builder()
                                      .userId(userId)
                                      .point(chargePoint)
